backend/requirements.txt
# AI-Powered Campaign Management Tool - Backend Dependencies

# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Data processing
pandas==2.1.3
numpy==1.25.2

# Validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP client for AI services
httpx==0.25.2
aiohttp==3.9.1

# Environment and configuration
python-dotenv==1.0.0
python-multipart==0.0.6

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# CORS and middleware
fastapi-cors==0.0.6

# File handling
openpyxl==3.1.2
xlsxwriter==3.1.9

# Date and time utilities
python-dateutil==2.8.2

# Logging
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# AI/ML (optional - for advanced features)
openai==1.3.7
langchain==0.0.350
langchain-openai==0.0.2

# Monitoring and observability
prometheus-client==0.19.0
